bot/requirements.txt
aiomysql>=0.2.0
aiohttp>=3.10.5
discord.py>=2.4.0
python-dotenv>=1.0.0